# Speed up generated decoder compilation times