--- a/rustler_codegen/src/context.rs
+++ b/rustler_codegen/src/context.rs
@@ -122,15 +122,20 @@ impl<'a> Context<'a> {
 
     pub fn escape_ident_with_index(ident_str: &str, index: usize, infix: &str) -> Ident {
         Ident::new(
-            &format!("RUSTLER_{}_field_{}_{}", infix, index, Self::remove_raw(ident_str)),
-            Span::call_site()
+            &format!(
+                "RUSTLER_{}_field_{}_{}",
+                infix,
+                index,
+                Self::remove_raw(ident_str)
+            ),
+            Span::call_site(),
         )
     }
 
     pub fn escape_ident(ident_str: &str, infix: &str) -> Ident {
         Ident::new(
             &format!("RUSTLER_{}_field_{}", infix, Self::remove_raw(ident_str)),
-            Span::call_site()
+            Span::call_site(),
         )
     }