--- a/rustler_codegen/src/tuple.rs
+++ b/rustler_codegen/src/tuple.rs
@@ -1,6 +1,6 @@
-use proc_macro2::TokenStream;
+use proc_macro2::{Span, TokenStream};
 
-use syn::{self, Field, Index};
+use syn::{self, Field, Ident, Index};
 
 use super::context::Context;
 
@@ -35,9 +35,10 @@ pub fn transcoder_decorator(ast: &syn::DeriveInput) -> TokenStream {
 fn gen_decoder(ctx: &Context, fields: &[&Field]) -> TokenStream {
     let struct_type = &ctx.ident_with_lifetime;
     let struct_name = ctx.ident;
+    let struct_name_str = struct_name.to_string();
 
     // Make a decoder for each of the fields in the struct.
-    let field_defs: Vec<TokenStream> = fields
+    let (assignments, field_defs): (Vec<TokenStream>, Vec<TokenStream>) = fields
         .iter()
         .enumerate()
         .map(|(index, field)| {
@@ -47,36 +48,39 @@ fn gen_decoder(ctx: &Context, fields: &[&Field]) -> TokenStream {
             } else {
                 index.to_string()
             };
-            let error_message = format!(
-                "Could not decode field {} on {}",
-                pos_in_struct,
-                struct_name.to_string()
+
+            let variable = Ident::new(
+                &format!("RUSTLER_TUPLE_field_{}", pos_in_struct),
+                Span::call_site(),
             );
 
-            let decoder = quote! {
-                match ::rustler::Decoder::decode(terms[#index]) {
-                    Err(_) => return Err(::rustler::Error::RaiseTerm(Box::new(#error_message))),
-                    Ok(value) => value
+            let assignment = quote! {
+                let #variable = try_decode_index(&terms, #pos_in_struct, #index)?;
+            };
+
+            let field_def = match ident {
+                None => quote! { #variable },
+                Some(ident) => {
+                    quote! { #ident: #variable }
                 }
             };
 
-            match ident {
-                None => quote! { #decoder },
-                Some(ident) => quote! { #ident: #decoder },
-            }
+            (assignment, field_def)
         })
-        .collect();
+        .unzip();
 
     let field_num = field_defs.len();
 
     // The implementation itself
     let construct = if ctx.is_tuple_struct {
         quote! {
-            #struct_name ( #(#field_defs),* )
+            #(#assignments);*
+            Ok(#struct_name ( #(#field_defs),* ))
         }
     } else {
         quote! {
-            #struct_name { #(#field_defs),* }
+            #(#assignments);*
+            Ok(#struct_name { #(#field_defs),* })
         }
     };
     let gen = quote! {
@@ -86,9 +90,18 @@ fn gen_decoder(ctx: &Context, fields: &[&Field]) -> TokenStream {
                 if terms.len() != #field_num {
                     return Err(::rustler::Error::BadArg);
                 }
-                Ok(
-                    #construct
-                )
+
+                fn try_decode_index<'a, T>(terms: &[::rustler::Term<'a>], pos_in_struct: &str, index: usize) -> Result<T, rustler::Error>
+                    where
+                        T: rustler::Decoder<'a>,
+                {
+                    match ::rustler::Decoder::decode(terms[index]) {
+                        Err(_) => Err(::rustler::Error::RaiseTerm(Box::new(
+                                    format!("Could not decode field {} on {}", pos_in_struct, #struct_name_str)))),
+                        Ok(value) => Ok(value)
+                    }
+                }
+                #construct
             }
         }
     };