This pull request speeds up compilation by extracting common code from field assignments into helper functions, and by preferring local variables for initial assignment instead of assigning struct fields directly. See #299 for the corresponding issue.

Fix #299.