@evnu 2020-01-29
Some benchmark results from running `meta_benchmark.sh` in https://github.com/evnu/rusty/tree/master/native/benchmark_compile.

```
~/tools/rusty/native/benchmark_compile(master ✗) ./meta_benchmark.sh
```

|Derive Type|master|299-speed-up-decoder-compilation|
|-|-|-|
|NifMap|![NifMap_master](https://user-images.githubusercontent.com/224554/73359332-e9a00380-42a0-11ea-815c-502228f2fbc0.png)|![NifMap_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73359333-e9a00380-42a0-11ea-8e54-c13948b85932.png)|
|NifStruct|![NifStruct_master](https://user-images.githubusercontent.com/224554/73359336-ea389a00-42a0-11ea-9e17-d49a8e7acf21.png)|![NifStruct_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73359338-ead13080-42a0-11ea-8b65-c9b976fd0be4.png)|
|NifRecord|![NifRecord_master](https://user-images.githubusercontent.com/224554/73359330-e9076d00-42a0-11ea-8499-c89fbe1c41b0.png)|![NifRecord_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73359331-e9a00380-42a0-11ea-993e-73516ea55844.png)|
|NifTuple|![NifTuple_master](https://user-images.githubusercontent.com/224554/73359334-ea389a00-42a0-11ea-8117-0958e160761b.png)|![NifTuple_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73359335-ea389a00-42a0-11ea-9159-61ed4d14b2d9.png)|

This looks very promising. I was not yet able to refactor `NifRecord` and `NifTuple` to achieve a similar speed up. The problem is that we need to introduce `let` assignments for the fields, but with tuple structs we do not have a field name available to name the let variable.