@evnu 2020-01-29
I managed to use `let` for tuple and records as well and squashed it into 2c36dee. Now, compilation is fast for `NifTuple` and `NifRecord` as well:

|Derive Type|master|299-speed-up-decoder-compilation|
|-|-|-|
|NifRecord|![NifRecord_master](https://user-images.githubusercontent.com/224554/73361419-3980c980-42a5-11ea-8469-505ec00dacd6.png)|![NifRecord_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73361420-3a196000-42a5-11ea-939c-6c7adcc018e9.png)|
|NifTuple|![NifTuple_master](https://user-images.githubusercontent.com/224554/73361421-3a196000-42a5-11ea-8a2a-97e36a44fe3c.png)|![NifTuple_299-speed-up-decoder-compilation](https://user-images.githubusercontent.com/224554/73361422-3a196000-42a5-11ea-9969-13dcbafa41c6.png)|